<refactor> Board의 역할 분리 리팩토링
- ChessGame.java : 게임 로직을 담당
- ChessView.java : view 담당
